% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta-negative-binomial-distribution.R
\name{BetaNegBinom}
\alias{BetaNegBinom}
\alias{dbnbinom}
\alias{rbnbinom}
\title{Beta-Negative Binomial distribution}
\usage{
dbnbinom(x, size, alpha = 1, beta = 1, log = FALSE)

rbnbinom(n, size, alpha = 1, beta = 1)
}
\arguments{
\item{x}{vector of quantiles.}

\item{size, alpha, beta}{parameters.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}

\item{n}{number of observations. If \code{length(n) > 1},
the length is taken to be the number required.}
}
\description{
Probability mass function and random generation
for the Beta-binomial distribution.
}
\details{
Probability mass function
\deqn{
f(x) = \frac{\Gamma(r+k)}{k! \Gamma(r)}
       \frac{\mathrm{B}(\alpha+r, \beta+k)}{\mathrm{B}(\alpha, \beta)}
}{
f(x) = gamma(r+k)/(k! gamma(r)) * beta(alpha+r, beta+k)/beta(alpha, beta)
}
}
\seealso{
\code{\link{Beta}}, \code{\link{NegBinomial}}
}
\keyword{distribution}

