% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-laplace-distribution.R
\name{DiscreteLaplace}
\alias{DiscreteLaplace}
\alias{ddlaplace}
\alias{pdlaplace}
\alias{rdlaplace}
\title{Discrete Laplace distribution}
\usage{
ddlaplace(x, location, scale, log = FALSE)

pdlaplace(q, location, scale, lower.tail = TRUE, log.p = FALSE)

rdlaplace(n, location, scale)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{location}{location parameter.}

\item{scale}{scale parameter; \code{0 < scale < 1}.}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P[X \le x]}
otherwise, \eqn{P[X > x]}.}

\item{n}{number of observations. If \code{length(n) > 1},
the length is taken to be the number required.}
}
\description{
Probability mass, distribution function and random generation
for the discrete Laplace distribution parametrized by location and scale.
}
\details{
If \eqn{U \sim \mathrm{Geometric}(1-p)}{U ~ Geometric(1-p)} and
\eqn{V \sim \mathrm{Geometric}(1-p)}{V ~ Geometric(1-p)},
then \eqn{U-V \sim \mathrm{DiscreteLaplace}(p)}{U-V ~ DiscreteLaplace(p)},
where geometric distribution is related to discrete Laplace distribution
in similar way as exponential distribution is related to Laplace distribution.

Probability mass function

\deqn{
f(x) = \frac{1-p}{1+p} p^{|x-\mu|}
}{
f(x) = (1-p)/(1+p) * p^(|x-\mu|)
}

Cumulative distribution function

\deqn{
F(x) = \left\{\begin{array}{ll}
\frac{p^{-|x-\mu|}}{1+p} & x < 0 \\
1 - \frac{p^{|x-\mu|+1}}{1+p} & x \ge 0
\end{array}\right.
}{
F(x) = [if x < 0:] p^-floor(x-\mu))/(1+p) [else:] 1-(p^(floor(x-\mu)+1))/(1+p)
}
}
\examples{

p <- 0.45
x <- rdlaplace(1e5, 0, p)
xx <- seq(-200, 200, by = 1)
plot(prop.table(table(x)))
lines(xx, ddlaplace(xx, 0, p), col = "red")
hist(pdlaplace(x, 0, p))
plot(ecdf(x))
lines(xx, pdlaplace(xx, 0, p), col = "red", type = "s")

}
\references{
Inusah, S., & Kozubowski, T.J. (2006). A discrete analogue of the Laplace distribution.
Journal of statistical planning and inference, 136(3), 1090-1102.

Kotz, S., Kozubowski, T., & Podgorski, K. (2012).
The Laplace distribution and generalizations: a revisit with applications
to communications, economics, engineering, and finance.
Springer Science & Business Media.
}
\concept{Discrete}
\concept{Univariate}
\keyword{distribution}
