% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-weibull-distribution.R
\name{DiscrWeibull}
\alias{DiscrWeibull}
\alias{ddweibull}
\alias{pdweibull}
\alias{qdweibull}
\alias{rdweibull}
\title{Discrete Weibull distribution}
\usage{
ddweibull(x, q, beta, log = FALSE)

pdweibull(x, q, beta, lower.tail = TRUE, log.p = FALSE)

qdweibull(p, q, beta, lower.tail = TRUE, log.p = FALSE)

rdweibull(n, q, beta)
}
\arguments{
\item{x}{vector of quantiles.}

\item{q, beta}{parameters.}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P[X \le x]}
otherwise, \eqn{P[X > x]}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If \code{length(n) > 1},
the length is taken to be the number required.}
}
\description{
Density, distribution function, quantile function and random generation
for the Discrete Weibull distribution.
}
\details{
Probability mass function
\deqn{
f(x) = q^{x^\beta} - q^{(x+1)^\beta}
}{
f(x) = q^x^\beta - q^(x+1)^\beta
}

Cumulative distribution function
\deqn{
F(x) = 1-q^{(x+1)^\beta}
}{
F(x) = 1-q^(x+1)^\beta
}

Quantile function
\deqn{
F^{-1}(p) = \left \lceil{\left(\frac{\log(1-p)}{\log(q)}\right)^{1/\beta} - 1}\right \rceil
}{
F^-1(p) = ceiling((\log(1-p)/\log(q))^(1/\beta) - 1)
}
}
\examples{

x <- rdweibull(1e5, 0.32, 1)
xx <- seq(-2, 100, by = 1)
plot(prop.table(table(x)), type = "h")
lines(xx, ddweibull(xx, .32, 1), col = "red")

# Notice: distribution of F(X) is far from uniform:
hist(pdweibull(x, .32, 1), 50)

plot(ecdf(x))
lines(xx, pdweibull(xx, .32, 1), col = "red", lwd = 2)

}
\references{
Nakagawa, T. and Osaki, S. (1975). The Discrete Weibull Distribution.
IEEE Transactions on Reliability, R-24, 300-301.
}
\seealso{
\code{\link[stats]{Weibull}}
}
\keyword{distribution}

