% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraDistr-deprecated.R
\name{extraDistr-deprecated}
\alias{extraDistr-deprecated}
\alias{dnst}
\alias{pnst}
\alias{qnst}
\alias{rnst}
\title{Deprecated functions}
\usage{
dnst(x, df, mu = 0, sigma = 1, log = FALSE)

pnst(q, df, mu = 0, sigma = 1, lower.tail = TRUE, log.p = FALSE)

qnst(p, df, mu = 0, sigma = 1, lower.tail = TRUE, log.p = FALSE)

rnst(n, df, mu = 0, sigma = 1)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{df}{degrees of freedom (> 0, maybe non-integer). \code{df = Inf} is allowed.}

\item{mu}{vector of locations}

\item{sigma}{vector of positive valued scale parameters.}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P[X \le x]}
otherwise, \eqn{P[X > x]}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If \code{length(n) > 1},
the length is taken to be the number required.}
}
\description{
\emph{Warning}: the \code{dnst}, \code{pnst}, \code{qnst}, and \code{rnst}
functions are now deprecated, use \code{dlst}, \code{plst}, \code{qlst},
and \code{rlst} functions instead.
}
\seealso{
\link{LocationScaleT}
}
