% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gamma-poisson-distribution.R
\name{GammaPoiss}
\alias{GammaPoiss}
\alias{dgpois}
\alias{pgpois}
\alias{rgpois}
\title{Gamma-Poisson distribution}
\usage{
dgpois(x, alpha, beta, log = FALSE)

pgpois(x, alpha, beta, lower.tail = TRUE, log.p = FALSE)

rgpois(n, alpha, beta)
}
\arguments{
\item{x}{vector of quantiles.}

\item{alpha,beta}{parameters.}

\item{log,log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P[X \le x]}
                       otherwise, \eqn{P[X > x]}.}

\item{n}{number of observations. If \code{length(n) > 1},
the length is taken to be the number required.}

\item{p}{vector of probabilities.}
}
\description{
Density, distribution function, quantile function and random generation
for the Gamma-Poisson distribution.
}
\details{
Probability density function
\deqn{
f(x) = \frac{\Gamma(x+\beta) \alpha^x}{\Gamma(\beta) (1+\alpha)^{\beta+x} x!}
}{
f(x) = (\Gamma(x+\beta)*\alpha^x) / (\Gamma(\beta)*(1+\alpha)^(\beta+x) * x!)
}
}
\keyword{distribution}

