% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-beta-distribution.R
\name{GeneralBeta}
\alias{GeneralBeta}
\alias{dgbeta}
\alias{pgbeta}
\alias{qgbeta}
\alias{rgbeta}
\title{General beta distribution}
\usage{
dgbeta(x, shape1, shape2, min = 0, max = 1, log = FALSE)

pgbeta(x, shape1, shape2, min = 0, max = 1, lower.tail = TRUE,
  log.p = FALSE)

qgbeta(p, shape1, shape2, min = 0, max = 1, lower.tail = TRUE,
  log.p = FALSE)

rgbeta(n, shape1, shape2, min = 0, max = 1, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles.}

\item{shape1, shape2}{non-negative parameters of the Beta distribution.}

\item{min, max}{lower and upper bounds.}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P[X \leq x]},
otherwise, \eqn{P[X > x]}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If \code{length(n) > 1},
the length is taken to be the number required.}
}
\description{
General form of beta distribution with lower and upper bounds
denoted as \code{min} and \code{max}. By default \code{min=0}
and \code{max=1} what leads to standard beta distribution.
}
\seealso{
\code{\link{Beta}}
}
\keyword{distribution}

