% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpd-distribution.R
\name{GPD}
\alias{GPD}
\alias{dgpd}
\alias{pgpd}
\alias{qgpd}
\alias{rgpd}
\title{Generalized Pareto Distribution}
\usage{
dgpd(x, mu = 0, sigma = 1, xi = 0, log = FALSE)

pgpd(x, mu = 0, sigma = 1, xi = 0, lower.tail = TRUE, log.p = FALSE)

qgpd(p, mu = 0, sigma = 1, xi = 0, lower.tail = TRUE, log.p = FALSE)

rgpd(n, mu = 0, sigma = 1, xi = 0)
}
\arguments{
\item{x}{vector of quantiles.}

\item{mu, sigma, xi}{location, scale, and shape parameters.}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P[X \le x]}
otherwise, \eqn{P[X > x]}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If \code{length(n) > 1},
the length is taken to be the number required.}
}
\description{
Density, distribution function, quantile function and random generation
for the Generalized Pareto distribution.
}
\details{
Probability density function
\deqn{ f(x) = \left\{\begin{array}{ll}
(1+\xi z)^{-(\xi+1)/\xi} & \xi \neq 0 \\
\exp(-z)                 & \xi = 0
\end{array}\right.
}{
f(x) = [if \xi != 0:] (1+\xi*z)^{-(\xi+1)/\xi}
       [else:] exp(-z)
}

Cumulative distribution function
\deqn{ F(x) = \left\{\begin{array}{ll}
1-(1+\xi z)^{-1/\xi} & \xi \neq 0 \\
1-\exp(-z)           & \xi = 0
\end{array}\right.
}{
F(x) = [if \xi != 0:] 1-(1+\xi*z)^{-1/\xi}
       [else:] 1-exp(-z)
}

Quantile function
\deqn{ F^{-1}(x) = \left\{\begin{array}{ll}
\mu + \sigma \frac{(1-p)^{-\xi}-1}{\xi} & \xi \neq 0 \\
\mu - \sigma \log(1-p)            & \xi = 0
\end{array}\right.
}{
F^-1(x) = [if \xi != 0:] \mu + \sigma * ((1-p)^{-\xi}-1)/\xi
          [else:] \mu - \sigma * log(1-p)
}

where \eqn{ z = \frac{x-\mu}{\sigma} }{ z = (x-\mu)/\sigma }.
}
\references{
Coles, S. (2001). An Introduction to Statistical Modeling of Extreme Values. Springer.
}
\keyword{distribution}

