% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gumbel-distribution.R
\name{Gumbel}
\alias{Gumbel}
\alias{dgumbel}
\alias{pgumbel}
\alias{qgumbel}
\alias{rgumbel}
\title{Gumbel distribution}
\usage{
dgumbel(x, mu = 0, sigma = 1, log = FALSE)

pgumbel(q, mu = 0, sigma = 1, lower.tail = TRUE, log.p = FALSE)

qgumbel(p, mu = 0, sigma = 1, lower.tail = TRUE, log.p = FALSE)

rgumbel(n, mu = 0, sigma = 1)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{mu, sigma}{location and scale parameters. Scale must be positive.}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P[X \le x]}
otherwise, \eqn{P[X > x]}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If \code{length(n) > 1},
the length is taken to be the number required.}
}
\description{
Density, distribution function, quantile function and random generation
for the Gumbel distribution.
}
\details{
Probability density function
\deqn{
f(x) = \frac{1}{\sigma} \exp(-(z+\exp(-z)))
}{
f(x) = 1/\sigma * exp(-(z+exp(-z)))
}

Cumulative distribution function
\deqn{
F(x) = \exp(-\exp(-z))
}{
F(x) = exp(-exp(-z))
}

Quantile function
\deqn{
F^{-1}(p) = \mu - \sigma \log(-\log(p))
}{
F^-1(p) = \mu - \sigma * log(-log(p))
}

where \eqn{ z = \frac{x-\mu}{\sigma} }{ z = (x-\mu)/\sigma }.
}
\examples{

x <- rgumbel(1e5, 5, 2)
xx <- seq(0, 1000, by = 0.1)
hist(x, 100, freq = FALSE)
lines(xx, dgumbel(xx, 5, 2), col = "red")
hist(pgumbel(x, 5, 2))
plot(ecdf(x))
lines(xx, pgumbel(xx, 5, 2), col = "red", lwd = 2)

}
\references{
Bury, K. (1999). Statistical Distributions in Engineering.
Cambridge University Press.
}
\keyword{distribution}

