% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huber-distribution.R
\name{Huber}
\alias{Huber}
\alias{dhuber}
\alias{phuber}
\alias{qhuber}
\alias{rhuber}
\title{"Huber Density" distribution}
\usage{
dhuber(x, mu = 0, sigma = 1, epsilon = 1.345, log = FALSE)

phuber(q, mu = 0, sigma = 1, epsilon = 1.345, lower.tail = TRUE,
  log.p = FALSE)

qhuber(p, mu = 0, sigma = 1, epsilon = 1.345, lower.tail = TRUE,
  log.p = FALSE)

rhuber(n, mu = 0, sigma = 1, epsilon = 1.345)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{mu, sigma, epsilon}{location, and scale, and shape parameters.
Scale and shape must be positive.}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P[X \le x]}
otherwise, \eqn{P[X > x]}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If \code{length(n) > 1},
the length is taken to be the number required.}
}
\description{
Density, distribution function, quantile function and random generation
for the "Huber density" distribution.
}
\examples{

x <- rhuber(1e5, 5, 2, 3)
xx <- seq(-20, 20, by = 0.1)
hist(x, 100, freq = FALSE)
lines(xx, dhuber(xx, 5, 2, 3), col = "red")
hist(phuber(x, 5, 2, 3))
plot(ecdf(x))
lines(xx, phuber(xx, 5, 2, 3), col = "red", lwd = 2)

}
\keyword{distribution}

