% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logarithmic-series-distribution.R
\name{LogSeries}
\alias{LogSeries}
\alias{dlgser}
\alias{plgser}
\alias{rlgser}
\title{Logarythmic Series distribution}
\usage{
dlgser(x, theta, log = FALSE)

plgser(x, theta, lower.tail = TRUE, log.p = FALSE)

rlgser(n, theta)
}
\arguments{
\item{x}{matrix of quantiles.}

\item{theta}{vector; concentration parameter.}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P[X \le x]}
otherwise, \eqn{P[X > x]}.}

\item{n}{number of observations. If \code{length(n) > 1},
the length is taken to be the number required.}
}
\description{
Density function, cumulative distribution function and random generation
for the Logarythmic Series distribution.
}
\details{
Probability mass function
\deqn{
f(x) = \frac{-1}{\log(1-\theta)} \frac{\theta^x}{x}
}{
f(x) = (-1/log(1-\theta)*\theta^x) / x
}

Cumulative distribution function
\deqn{
F(x) = \frac{-1}{\log(1-\theta)} \sum_{k=1}^x \frac{\theta^x}{x}
}{
F(x) = -1/log(1-\theta) * sum((\theta^x)/x)
}

Random generation is done using algorithm described in Krishnamoorthy (2006).
}
\references{
Krishnamoorthy, K. (2006). Handbook of Statistical Distributions
with Applications. Chapman & Hall/CRC

Forbes, C., Evans, M. Hastings, N., & Peacock, B. (2011).
Statistical Distributions. John Wiley & Sons.
}

