% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nakagami-distribution.R
\name{Nakagami}
\alias{Nakagami}
\alias{dnaka}
\alias{pnaka}
\alias{qnaka}
\alias{rnaka}
\title{Nakagami distribution}
\usage{
dnaka(x, m, w, log = FALSE)

pnaka(q, m, w, lower.tail = TRUE, log.p = FALSE)

qnaka(p, m, w, lower.tail = TRUE, log.p = FALSE)

rnaka(n, m, w)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{m}{positive valued shape parameter.}

\item{w}{positive valued scale parameter.}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P[X \le x]}
otherwise, \eqn{P[X > x]}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If \code{length(n) > 1},
the length is taken to be the number required.}
}
\description{
Density, distribution function, quantile function and random generation
for the Nakagami distribution.
}
\details{
The Nakagami distribution (Nakagami, 1960) with shape \eqn{m} and scale
\eqn{\omega} has density
\deqn{
f(x) = \frac{2m^m}{\Gamma(m)\omega^m} x^{2m-1} \exp\left(-\frac{m}{\omega} x^2\right)
}{
f(x) = (2*m^m / (gamma(m)*\omega^m)) * x^(2m-1) * exp(-m/\omega * x^2)
}
for \eqn{x \ge 0}, \eqn{m > 0} and \eqn{\omega > 0}.

If \eqn{Y} is Gamma distributed with shape \eqn{m} and rate \eqn{m/\omega},
then \eqn{X = \sqrt{Y}} is Nakagami distributed with shape \eqn{m} and
scale \eqn{\omega}.
}
\examples{

x <- rnaka(1e5, m = 2, w = 1)
hist(x, 100, freq = FALSE)
curve(dnaka(x, m = 2, w = 1), 0, 3, col = "red", add = TRUE)
hist(pnaka(x, m = 2, w = 1))
plot(ecdf(x))
curve(pnaka(x, m = 2, w = 1), 0, 3, col = "red", lwd = 2, add = TRUE)

}
\references{
Nakagami, M. (1960). The m-Distribution - A General Formula of
Intensity Distribution of Rapid Fading. In Statistical Methods in
Radio Wave Propagation, edited by W. C. Hoffman, 3-36. Pergamon Press.
}
\seealso{
\code{\link[stats]{GammaDist}}
}
\concept{Continuous}
\concept{Univariate}
\keyword{distribution}
