% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/poisson-binomial-distribution.R
\name{PoisBinom}
\alias{PoisBinom}
\alias{dpbinom}
\alias{ppbinom}
\alias{rpbinom}
\title{Poisson-binomial distribution}
\usage{
dpbinom(x, prob, log = FALSE, method = c("MC", "PA", "NA", "BA", "RA",
  "RNA"), nsim = 10000L)

ppbinom(x, prob, lower.tail = TRUE, log.p = FALSE, method = c("MC", "PA",
  "NA", "BA", "RA", "RNA"), nsim = 10000L)

rpbinom(n, prob)
}
\arguments{
\item{x}{vector of quantiles.}

\item{prob}{parameters.}

\item{log,log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{method}{method for approximating density and distribution function;
                       "MC" for Monte Carlo simulation, "PA" for poisson approximation,
                       "BA" for Binomial approximation, "NA" for Normal approximation,
                       "RNA" for refined Normal approximation.}

\item{nsim}{number of samples in Monte Carlo simulation for calculating
density and distribution function; the higher is more precise
but slower.}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P[X \le x]}
otherwise, \eqn{P[X > x]}.}

\item{n}{number of observations. If \code{length(n) > 1},
the length is taken to be the number required.}

\item{p}{vector of probabilities.}
}
\description{
Probability mass function, distribution function and random generation
for the Poisson-binomial distribution.
}
\details{
There are multiple ways of estimating probability mass function and distribution function
of Poisson-Binomial, however direct methods are numerically unstable. Multiple indirect
and approximate methods were proposed (Chen and Liu, 1997, Hong, 2013, Volkowa, 1996,
Choi and Xia, 2002). This function enables to use Monte Carlo simulation and
approximation-based methods. Poisson-Binomial can be approximated using Poisson
distribution by Le Cam theorem, however Binomial approximation is generally better
(Chen and Liu, 2002). Normal approximation or its correction proposed by Volkova (1996)
are also possible. Other methods are available in \pkg{poibin} package by Yili Hong.
}
\references{
Chen, L.H.Y. (1974). On the Convergence of Poisson Binomial to Poisson
Distributions. The Annals of Probability, 2(1), 178-180.

Chen, S.X. and Liu, J.S. (1997). Statistical applications of the Poisson-Binomial
and conditional Bernoulli distributions. Statistica Sinica 7, 875-892.

Chen, S.X. (1993). Poisson-Binomial distribution, conditional Bernoulli distribution
and maximum entropy. Technical Report. Department of Statistics, Harvard University.

Chen, X.H., Dempster, A.P. and Liu, J.S. (1994). Weighted finite population
sampling to maximize entropy. Biometrika 81, 457-469.

Wang, Y.H. (1993). On the number of successes in independent trials.
Statistica Sinica 3(2): 295-312.

Hong, Y. (2013). On computing the distribution function for the Poisson
binomial distribution. Computational Statistics & Data Analysis, 59, 41-51.

Volkova, A. Y. (1996). A refinement of the central limit theorem for sums
of independent random indicators. Theory of Probability and its Applications 40, 791-794.

Choi, K.P. and Xia, A. (2002). Approximating the number of successes in independent trials:
Binomial versus Poisson. The Annals of Applied Probability, 14(4), 1139-1148.
}
\keyword{distribution}

