% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rademacher-distribution.R
\name{Rademacher}
\alias{Rademacher}
\alias{rsign}
\title{Random generation from Rademacher distribution}
\usage{
rsign(n)
}
\arguments{
\item{n}{number of observations. If \code{length(n) > 1},
the length is taken to be the number required.}
}
\description{
Random generation for the Rademacher distribution (values -1 and +1
with equal probability).
}
\keyword{distribution}
