% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaled-inverse-chi-squared-distribution.R
\name{ScInvChiSq}
\alias{ScInvChiSq}
\alias{dsinvchisq}
\alias{psinvchisq}
\alias{rsinvchisq}
\title{Inverse-chi-squared distribution}
\usage{
dsinvchisq(x, nu, tau = 1, log = FALSE)

psinvchisq(x, nu, tau = 1, lower.tail = TRUE, log.p = FALSE)

rsinvchisq(n, nu, tau = 1)
}
\arguments{
\item{x}{vector of quantiles.}

\item{nu, tau}{shape and scale parameters.}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P[X \le x]}
otherwise, \eqn{P[X > x]}.}

\item{n}{number of observations. If \code{length(n) > 1},
the length is taken to be the number required.}

\item{p}{vector of probabilities.}
}
\description{
Density, distribution function and random generation
for the Inverse-chi-squared distribution.
}
\details{
Probability mass function
\deqn{
f(x) = \frac{2^{-\nu/2}}{\Gamma(\nu/2)} x^{-\nu/2-1} e^{-1/2x}
}{
f(x) = (2^(-\nu/2))/\Gamma(\nu/2) * x^(-\nu/2-1) * exp(-1/(2*x))
}

Cumulative distribution function
\deqn{
F(x) = \frac{\gamma(\nu/2, 1/2x)}{\Gamma(\nu/2)}
}{
F(x) = \gamma(\nu/2, 1/(2*x)) / \Gamma(\nu/2)
}
}
\seealso{
\code{\link{Chisquare}}
}
\keyword{distribution}

