% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/beta-binomial-distribution.R
\name{BetaBinom}
\alias{BetaBinom}
\alias{dbbinom}
\alias{pbbinom}
\alias{rbbinom}
\title{Beta-Binomial distribution}
\usage{
dbbinom(x, size, alpha = 1, beta = 1, log = FALSE)

pbbinom(x, size, alpha = 1, beta = 1, lower.tail = TRUE, log.p = FALSE)

rbbinom(n, size, alpha = 1, beta = 1)
}
\arguments{
\item{x}{vector of quantiles.}

\item{size,alpha,beta}{parameters.}

\item{log,log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P[X \le x]}
                       otherwise, \eqn{P[X > x]}.}

\item{n}{number of observations. If \code{length(n) > 1},
the length is taken to be the number required.}

\item{p}{vector of probabilities.}
}
\description{
Density, distribution function and random generation
for the Beta-binomial distribution.
}
\details{
Probability mass function
\deqn{
f(x) = {n \choose x} \frac{\mathrm{B}(x+\alpha, n-x+\beta)}{\mathrm{B}(\alpha, \beta)}
}{
f(x) = choose(n, x) * (beta(x+alpha, n-x+beta)) / (beta(alpha, beta))
}
}
\seealso{
\code{\link{Beta}}, \code{\link{Binomial}}
}
\keyword{distribution}

