% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate-normal-distribution.R
\name{BivNormal}
\alias{BivNormal}
\alias{dbvnorm}
\alias{rbvnorm}
\title{Bivariate Normal distribution}
\usage{
dbvnorm(x, y, mu1 = 0, mu2 = mu1, sigma1 = 1, sigma2 = sigma1,
  rho = 0, log = FALSE)

rbvnorm(n, mu1 = 0, mu2 = mu1, sigma1 = 1, sigma2 = sigma1, rho = 0)
}
\arguments{
\item{x, y}{vectors of quantiles.}

\item{mu1, mu2, sigma1, sigma2, rho}{location, scale and correlation parameters.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}

\item{n}{number of observations. If \code{length(n) > 1},
the length is taken to be the number required.}
}
\description{
Density, distribution function and random generation
for the Bivariate Normal distribution.
}
\details{
Probability density function
\deqn{
f(x) = \frac{1}{2\pi\sqrt{1-\rho^2}\sigma_1\sigma_2}
       \exp\left(-\frac{1}{2(1-\rho^2)} (z_1^2 - 2\rho z_1 z_2 + z_2^2)\right)
}{
f(x) = 1/(2*pi*sqrt(1-rho^2)*sigma1*sigma2) * exp(-(1/(2*(1-rho^2)*(z1^2 - 2*rho*z1*z2 * z2^2))))
}

where \eqn{
z_1 = \frac{x_1 - \mu_1}{\sigma_1}
}{
z1 = (x1 - \mu1)/\sigma1
} and \eqn{
z_2 = \frac{x_2 - \mu_2}{\sigma_2}
}{
z2 = (x2 - \mu2)/\sigma2
}.

For multivariate Normal and t distributions check \pkg{mvtnorm} package.
}
\references{
Krishnamoorthy, K. (2006). Handbook of Statistical Distributions
with Applications. Chapman & Hall/CRC
}
\keyword{distribution}

