% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dirichlet-distribution.R
\name{Dirichlet}
\alias{Dirichlet}
\alias{ddirichlet}
\alias{pdirichlet}
\alias{rdirichlet}
\title{Dirichlet distribution}
\usage{
ddirichlet(x, alpha, log = FALSE)

pdirichlet(x, alpha, lower.tail = TRUE, log.p = FALSE, nsim = 10000L)

rdirichlet(n, alpha)
}
\arguments{
\item{x}{matrix of quantiles.}

\item{alpha}{vector; concentration parameter.}

\item{log,log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P[X \le x]}
otherwise, \eqn{P[X > x]}.}

\item{nsim}{number of samples in Monte Carlo simulation for calculating
                       cumulative distribution function; the higher is more precise
                       but slower.}

\item{n}{number of observations. If \code{length(n) > 1},
the length is taken to be the number required.}
}
\description{
Density function, cumulative distribution function and random generation
for the Dirichlet distribution.
}
\details{
Probability density function
\deqn{
f(x) = \frac{\Gamma(\sum_k \alpha_k)}{\prod_k \Gamma(\alpha_k)} \prod_k x_k^{k-1}
}{
f(x) = \Gamma(sum(\alpha[k])) / prod(\Gamma(\alpha[k])) * prod(x[k]^{k-1})
}

Cumulative distribution function is approximated using Monte Carlo simulation.
}
\references{
Devroye, L. (1986). Non-Uniform Random Variate Generation. Springer-Verlag.

\url{http://stats.stackexchange.com/a/57313/35989}
}

