% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma-poisson-distribution.R
\name{GammaPoiss}
\alias{GammaPoiss}
\alias{dgpois}
\alias{rgpois}
\title{Gamma-Poisson distribution}
\usage{
dgpois(x, shape, rate, scale = 1/rate, log = FALSE)

rgpois(n, shape, rate, scale = 1/rate)
}
\arguments{
\item{x}{vector of quantiles.}

\item{shape, scale}{shape and scale parameters. Must be positive,
scale strictly.}

\item{rate}{an alternative way to specify the scale.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}

\item{n}{number of observations. If \code{length(n) > 1},
the length is taken to be the number required.}
}
\description{
Probability mass function and random generation
for the Gamma-Poisson distribution.
}
\details{
Gamma-Poisson distribution arises as a continuous mixture of
Poisson distributions, where where the mixing distribution
of the Poisson rate \eqn{\lambda} is a gamma distribution.
When \eqn{X \sim \mathrm{Poisson}(\lambda)}{X ~ Poisson(\lambda)}
and \eqn{\lambda \sim \mathrm{Gamma}(\alpha, \beta)}{\lambda ~ Gamma(\alpha, \beta)}, then \eqn{X \sim \mathrm{GammaPoisson}(\alpha, \beta)}{X ~ Gamma-Poisson(\alpha, \beta)}.

Probability density function (parametrized by scale)
\deqn{
f(x) = \frac{\Gamma(\alpha+x)}{x! \Gamma(\alpha)} p^k (1-p)^\alpha
}{
f(x) = \Gamma(\alpha+x) / (x!*\Gamma(\alpha)) * p^x * (1-p)^\alpha
}

where \eqn{p = \frac{\beta}{1+\beta}}{p = \beta/(1+\beta)}.
}
\seealso{
\code{\link{Gamma}}, \code{\link{Poisson}}
}
\keyword{distribution}

