% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncated-negative-binomial-distribution.R
\name{TruncNegBinom}
\alias{TruncNegBinom}
\alias{dtnbinom}
\alias{TruncNB}
\alias{ptnbinom}
\alias{qtnbinom}
\alias{rtnbinom}
\title{Truncated negative binomial distribution}
\usage{
dtnbinom(x, size, prob, mu, a = -Inf, b = Inf, log = FALSE)

ptnbinom(
  q,
  size,
  prob,
  mu,
  a = -Inf,
  b = Inf,
  lower.tail = TRUE,
  log.p = FALSE
)

qtnbinom(
  p,
  size,
  prob,
  mu,
  a = -Inf,
  b = Inf,
  lower.tail = TRUE,
  log.p = FALSE
)

rtnbinom(n, size, prob, mu, a = -Inf, b = Inf)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{size}{target for number of successful trials, or dispersion
parameter (the shape parameter of the gamma mixing
distribution). Must be strictly positive, need not be
integer.}

\item{prob}{probability of success in each trial. \code{0 < prob <= 1}.}

\item{mu}{alternative parameterization via mean}

\item{a, b}{lower and upper truncation points (\code{a < x <= b}).}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P[X \le x]}
otherwise, \eqn{P[X > x]}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If \code{length(n) > 1},
the length is taken to be the number required.}
}
\description{
Density, distribution function, quantile function and random generation
for the truncated negative binomial distribution.
}
\examples{

x <- rtnbinom(1e5, 14, 16)
xx <- seq(-1, 50)
plot(prop.table(table(x)))
lines(xx, dtpois(xx, 14, 16), col = "red")
hist(ptpois(x, 14, 16))

xx <- seq(0, 50, by = 0.01)
plot(ecdf(x))
lines(xx, ptpois(xx, 14, 16), col = "red", lwd = 2)

uu <- seq(0, 1, by = 0.001)
lines(qtpois(uu, 14, 16), uu, col = "blue", lty = 2)

# Zero-truncated Poisson

x <- rtpois(1e5, 5, 0)
xx <- seq(-1, 50)
plot(prop.table(table(x)))
lines(xx, dtpois(xx, 5, 0), col = "red")
hist(ptpois(x, 5, 0))

xx <- seq(0, 50, by = 0.01)
plot(ecdf(x))
lines(xx, ptpois(xx, 5, 0), col = "red", lwd = 2)
lines(qtpois(uu, 5, 0), uu, col = "blue", lty = 2)

}
\references{
Hilbe, J. (2011). Censored and truncated count models. In 
*Negative Binomial Regression* (pp. 387-406). Cambridge: Cambridge University 
Press. \url{https://doi.org/10.1017/CBO9780511973420.013}
}
\seealso{
\code{\link[stats]{NegBinomial}}
}
\concept{Discrete}
\concept{Univariate}
\keyword{distribution}
